append.to.last.tryCatchLog.result {tryCatchLog}	R Documentation

Appends a new log entry to the stored logging output of the last call to tryCatchLog or tryLog

Description

You can get the last logging output by calling last.tryCatchLog.result.

Usage

append.to.last.tryCatchLog.result(new.log.entry)

Arguments

new.log.entry
the new log entry (a data.frame created with link{build.log.entry})

Details

THIS FUNCTION IS USED ONLY PACKAGE INTERNALLY!

Value

the complete logging result of the last call to tryCatchLog or tryLog as data.frame

Note

THIS IS A PACKAGE INTERNAL FUNCTION AND THEREFORE NOT EXPORTED.

See Also

last.tryCatchLog.result, reset.last.tryCatchLog.result,

[Package tryCatchLog version 1.3.3 Index]