build.log.entry {tryCatchLog}	R Documentation

Creates a log entry as a single data.frame row containing all relevant logging information in columns

Description

The serverity level should correspond to the condition class.

Usage

build.log.entry(
  timestamp,
  severity,
  msg.text,
  execution.context.msg,
  call.stack,
  dump.file.name,
  omit.call.stack.items = 0
)

Arguments

timestamp
logging timestamp as POSIXct (normally by calling Sys.time)

severity
severity level of the log entry ((ERROR, WARN, INFO etc.)

msg.text
Logging message (e. g. error message)

execution.context.msg
a text identifier (eg. the PID or a variable value) that will be appended to msg.text for catched conditions. Must be a character or an error is thrown.

call.stack
a call stack created by sys.calls

dump.file.name
name of the created dump file (leave empty if the tryCatchLog argument write.error.dump.file is FALSE)

omit.call.stack.items
the number of stack trace items to ignore (= last x calls) in the passed call.stack since they are caused by using tryCatchLog

Value

An object of class tryCatchLog.log.entry and data.frame and the following columns:

timestamp - creation date and time of the logging entry

severity - the serverity level of the log entry (ERROR, WARN, INFO etc.)

msg.text - the message text of the log entry

compact.stack.trace - the short stack trace containing only entries with source code references down to line of code that has thrown the condition

full.stack.trace - the full stack trace with all calls down to the line of code that has thrown the condition (including calls to R internal functions and other functions even when the source code in not available).

dump.file.name - name of the created dump file (if any)

Note

THIS IS A PACKAGE INTERNAL FUNCTION AND THEREFORE NOT EXPORTED.

See Also

last.tryCatchLog.result build.log.output

[Package tryCatchLog version 1.3.3 Index]