get.pretty.option.value {tryCatchLog}	R Documentation

gets the current value of an option as key/value string

Description

The data type is also indicated if an option is set (since a wrong data type may cause problems). If an option is not set "(not set)" is shown as value.

Usage

get.pretty.option.value(option.name)

Arguments

option.name
Name of the option (as character)

Details

THIS IS AN INTERNAL PRIVATE FUNCTION OF THE PACKAGE.

Value

The option as key/value string in one line

See Also

get.pretty.tryCatchLog.options

Examples

## Not run: 
tryCatchLog:::get.pretty.option.value("warn")
# [1] "Option warn = 0 (double)" 
## End(Not run)

[Package tryCatchLog version 1.3.3 Index]