get.pretty.tryCatchLog.options {tryCatchLog}	R Documentation

Gets the current option values of all options supported by the 'tryCatchLog' package

Description

This is a convenience function whose result can be used e. g. to log the current settings.

Usage

get.pretty.tryCatchLog.options()

Details

If an option is not set the string "(not set)" is shown as value.

The data type is also indicated if an option is set (since a wrong data type may cause problems).

Value

The current option settings as string (one per line as key/value pair), e. g.

Option tryCatchLog.write.error.dump.file = FALSE (logical)
Option tryCatchLog.write.error.folder = . (character)
Option tryCatchLog.silent.warnings = FALSE (logical)
Option tryCatchLog.silent.messages = (not set)

Examples

cat(get.pretty.tryCatchLog.options())  # "cat" does apply new line escape characters

[Package tryCatchLog version 1.3.3 Index]