is.duplicated.log.entry {tryCatchLog}	R Documentation

Check if a new log entry would be a duplicate of on an already existing log entry

Description

The log.entry is checked against the existing log entries from last.tryCatchLog.result using the following columns:

msg.text

full.stack.trace

Usage

is.duplicated.log.entry(log.entry)

Arguments

log.entry
A data.frame with the new log entry (exactly one row)

Value

TRUE if the log.entry is a duplicate, else FALSE

Note

Required function to fix issue #18 (https://github.com/aryoda/tryCatchLog/issues/18)

See Also

last.tryCatchLog.result, build.log.entry

[Package tryCatchLog version 1.3.3 Index]