is.package.available {tryCatchLog}	R Documentation

Checks if a package is installed and can be loaded

Description

Use this function to check for optional package dependencies within this package.

Usage

is.package.available(package.name)

Arguments

package.name
Name of the package (as string)

Details

This is a package-internal function!

See section ‘Good practice’ in '?.onAttach'.

Value

TRUE if the packages is installed, otherwise FALSE (invisible)

http://r-pkgs.had.co.nz/description.html

Examples

tryCatchLog:::is.package.available("tryCatchLog")  # must be TRUE :-)

[Package tryCatchLog version 1.3.3 Index]