is.windows {tryCatchLog}	R Documentation

Determines if R is running on a Windows operating system

Description

Throws a warning if an indication for Windows OS were found but the Windows OS cannot be recognized for sure (via a second different check).

Usage

is.windows()

Value

TRUE of running on a Windows OS else FALSE

Examples

is.windows()

[Package tryCatchLog version 1.3.3 Index]