reset.last.tryCatchLog.result {tryCatchLog}	R Documentation

Resets the stored logging output of the last call to tryCatchLog or tryLog to an empty list

Description

You can get the last logging output by calling last.tryCatchLog.result.

Usage

reset.last.tryCatchLog.result()

Value

invisible: TRUE

Note

THIS IS A PACKAGE INTERNAL FUNCTION AND THEREFORE NOT EXPORTED.

See Also

last.tryCatchLog.result, append.to.last.tryCatchLog.result,

[Package tryCatchLog version 1.3.3 Index]